src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile /usr/src/code/src/requirements.in
#
anyconfig==0.9.11
    # via kedro
appdirs==1.4.4
    # via black
argon2-cffi==20.1.0
    # via notebook
arrow==0.17.0
    # via jinja2-time
async-generator==1.10
    # via nbclient
attrs==20.3.0
    # via
    #   black
    #   jsonschema
    #   pytest
backcall==0.2.0
    # via ipython
binaryornot==0.4.4
    # via cookiecutter
black==v19.10b0
    # via -r /usr/src/code/src/requirements.in
bleach==3.3.0
    # via nbconvert
cachetools==4.2.1
    # via kedro
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via argon2-cffi
chardet==4.0.0
    # via
    #   binaryornot
    #   requests
click==7.1.2
    # via
    #   black
    #   cookiecutter
    #   kedro
    #   pip-tools
cookiecutter==1.6.0
    # via kedro
coverage==5.4
    # via pytest-cov
cycler==0.10.0
    # via matplotlib
decorator==4.4.2
    # via ipython
defusedxml==0.6.0
    # via nbconvert
entrypoints==0.3
    # via nbconvert
flake8==3.8.4
    # via -r /usr/src/code/src/requirements.in
fsspec==0.6.3
    # via kedro
future==0.18.2
    # via cookiecutter
h5py==3.1.0
    # via -r /usr/src/code/src/requirements.in
idna==2.10
    # via requests
ipykernel==5.5.0
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython-genutils==0.2.0
    # via
    #   jupyterlab
    #   nbformat
    #   notebook
    #   qtconsole
    #   traitlets
ipython==7.20.0
    # via
    #   -r /usr/src/code/src/requirements.in
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipywidgets==7.6.3
    # via jupyter
isort==4.3.21
    # via -r /usr/src/code/src/requirements.in
jedi==0.18.0
    # via ipython
jinja2-time==0.2.0
    # via cookiecutter
jinja2==2.11.3
    # via
    #   cookiecutter
    #   jinja2-time
    #   nbconvert
    #   notebook
jmespath==0.10.0
    # via kedro
joblib==1.0.1
    # via scikit-learn
jsonschema==3.2.0
    # via
    #   jupyterlab-launcher
    #   nbformat
jupyter-client==6.1.11
    # via
    #   -r /usr/src/code/src/requirements.in
    #   ipykernel
    #   jupyter-console
    #   kedro
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.2.0
    # via jupyter
jupyter-core==4.7.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyter==1.0.0
    # via -r /usr/src/code/src/requirements.in
jupyterlab-launcher==0.10.5
    # via jupyterlab
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.0
    # via ipywidgets
jupyterlab==0.31.1
    # via -r /usr/src/code/src/requirements.in
kedro==0.16.6
    # via -r /usr/src/code/src/requirements.in
kiwisolver==1.3.1
    # via matplotlib
markupsafe==1.1.1
    # via jinja2
matplotlib==3.3.4
    # via
    #   -r /usr/src/code/src/requirements.in
    #   missingno
    #   seaborn
mccabe==0.6.1
    # via flake8
missingno==0.4.2
    # via -r /usr/src/code/src/requirements.in
mistune==0.8.4
    # via nbconvert
more-itertools==8.7.0
    # via pytest
nbclient==0.5.2
    # via nbconvert
nbconvert==6.0.7
    # via
    #   jupyter
    #   notebook
nbformat==5.1.2
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   nbstripout
    #   notebook
nbstripout==0.3.3
    # via -r /usr/src/code/src/requirements.in
nest-asyncio==1.5.1
    # via nbclient
notebook==6.2.0
    # via
    #   jupyter
    #   jupyterlab
    #   jupyterlab-launcher
    #   widgetsnbextension
numexpr==2.7.2
    # via tables
numpy==1.20.1
    # via
    #   h5py
    #   matplotlib
    #   missingno
    #   numexpr
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   tables
packaging==20.9
    # via
    #   bleach
    #   pytest
pandas==1.2.2
    # via seaborn
pandocfilters==1.4.3
    # via nbconvert
parso==0.8.1
    # via jedi
pathspec==0.8.1
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.1.0
    # via matplotlib
pip-tools==5.5.0
    # via kedro
plotly==4.14.3
    # via -r /usr/src/code/src/requirements.in
pluggy==0.13.1
    # via
    #   kedro
    #   pytest
poyo==0.5.0
    # via cookiecutter
prometheus-client==0.9.0
    # via notebook
prompt-toolkit==3.0.16
    # via
    #   ipython
    #   jupyter-console
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
py==1.10.0
    # via pytest
pyasn1==0.4.8
    # via pysmb
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.2.0
    # via flake8
pygments==2.8.0
    # via
    #   ipython
    #   jupyter-console
    #   jupyterlab-pygments
    #   nbconvert
    #   qtconsole
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.17.3
    # via jsonschema
pysmb==1.2.6
    # via -r /usr/src/code/src/requirements.in
pytest-cov==2.11.1
    # via -r /usr/src/code/src/requirements.in
pytest-mock==1.13.0
    # via -r /usr/src/code/src/requirements.in
pytest==5.4.3
    # via
    #   -r /usr/src/code/src/requirements.in
    #   pytest-cov
    #   pytest-mock
python-dateutil==2.8.1
    # via
    #   arrow
    #   jupyter-client
    #   matplotlib
    #   pandas
python-json-logger==0.1.11
    # via kedro
pytz==2021.1
    # via pandas
pyyaml==5.4.1
    # via kedro
pyzmq==22.0.3
    # via
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.0.2
    # via jupyter
qtpy==1.9.0
    # via qtconsole
regex==2020.11.13
    # via black
requests==2.25.1
    # via cookiecutter
retrying==1.3.3
    # via plotly
scikit-learn==0.24.1
    # via sklearn
scipy==1.6.0
    # via
    #   -r /usr/src/code/src/requirements.in
    #   missingno
    #   scikit-learn
    #   seaborn
seaborn==0.11.1
    # via
    #   -r /usr/src/code/src/requirements.in
    #   missingno
send2trash==1.5.0
    # via notebook
six==1.15.0
    # via
    #   argon2-cffi
    #   bleach
    #   cycler
    #   jsonschema
    #   plotly
    #   python-dateutil
    #   retrying
sklearn==0.0
    # via -r /usr/src/code/src/requirements.in
tables==3.6.1
    # via -r /usr/src/code/src/requirements.in
terminado==0.9.2
    # via notebook
testpath==0.4.4
    # via nbconvert
threadpoolctl==2.1.0
    # via scikit-learn
toml==0.10.2
    # via black
toposort==1.6
    # via kedro
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tqdm==4.56.0
    # via -r /usr/src/code/src/requirements.in
traitlets==5.0.5
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
typed-ast==1.4.2
    # via black
urllib3==1.26.3
    # via requests
wcwidth==0.2.5
    # via
    #   prompt-toolkit
    #   pytest
webencodings==0.5.1
    # via bleach
wheel==0.32.2
    # via -r /usr/src/code/src/requirements.in
whichcraft==0.6.1
    # via cookiecutter
widgetsnbextension==3.5.1
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
